requirements.txt
requests
pandas
python-dotenv
groq
flask